# Improve solve performance by caching possibility tables