--- a/README.md
+++ b/README.md
@@ -21,8 +21,8 @@ Current solve benchmarks:
 
 | Puzzle | Solved | Time (ms) | Iterations | Guesses |
 | ------------ | ------------ | ------------ | ------------ | ------------ |
-| Easy | True | 5300 | 8 | 0 |
-| Medium | True | 10427 | 11 | 0 |
-| Hard | False | 4993 | 4 | 0 |
-| Expert | False | 5808 | 4 | 0 |
+| Easy | True | 1316| 7 | 0 |
+| Medium | True | 1429| 10 | 0 |
+| Hard | False | 667 | 4 | 0 |
+| Expert | False | 547| 4 | 0 |
 | Evil | - | - | - | - |