… src/solve/utils.h