Add a table of possibilities in the form of a bitmap for each position. When we solve a position, we can remove that number from all of the possibilities in that row, column, and box. This allows us to do direct lookups instead of iterating to find the possibilities at each position.

The solve logic is pretty much unchanged at this point; but it runs MUCH faster with this optimization.

Before:

| Puzzle | Solved | Time (ms) | Iterations | Guesses |
| ------------ | ------------ | ------------ | ------------ | ------------ |
| Easy | True | 5300 | 8 | 0 |
| Medium | True | 10427 | 11 | 0 |
| Hard | False | 4993 | 4 | 0 |
| Expert | False | 5808 | 4 | 0 |
| Evil | - | - | - | - |

After:

| Puzzle | Solved | Time (ms) | Iterations | Guesses |
| ------------ | ------------ | ------------ | ------------ | ------------ |
| Easy | True | 1316| 7 | 0 |
| Medium | True | 1429| 10 | 0 |
| Hard | False | 667 | 4 | 0 |
| Expert | False | 547| 4 | 0 |
| Evil | - | - | - | - |